# Don't add fake docstring node everytime (once is enough)