--- a/jedi/evaluate/compiled/fake.py
+++ b/jedi/evaluate/compiled/fake.py
@@ -8,6 +8,7 @@ import os
 import inspect
 
 from jedi._compatibility import is_py3, builtins, unicode
+from jedi.cache import memoize_function
 from jedi.parser import Parser, load_grammar
 from jedi.parser import tree as pt
 from jedi.evaluate.helpers import FakeName
@@ -97,6 +98,7 @@ def _faked(module, obj, name):
             return search_scope(cls, name)
 
 
+@memoize_function
 def get_faked(module, obj, name=None):
     obj = obj.__class__ if is_class_instance(obj) else obj
     result = _faked(module, obj, name)