If the node is added everytime to ``obj``, we end up with a lots of redundant ``docstr_node`` that slow down Script.get_completion (and possibly other parts of the public API) when called multiple times within the same process.

I am using ``repr`` to check for node equality. Let me know if there is a better way.

Fix #591

For reference:

- **before the fix** the **first call** to jedi lasts **0.048 seconds** while the **last call** lasts **0.587 seconds**
- **after the fix**, the **first call** to jedi lasts **0.043 seconds** while the **last call** lasts **0.025 seconds**

Here is the output of the test script after the fix:

```
<object object at 0x7f5842a5f0e0> speed: init 7.939338684082031e-05
<object object at 0x7f5842a5f0e0> speed: completions start 0.00012803077697753906
<object object at 0x7f5842a5f0e0> speed: parsed 0.001271963119506836
<object object at 0x7f5842a5f0e0> speed: completions end 0.04212450981140137
         48193 function calls (47004 primitive calls) in 0.043 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
  493/214    0.005    0.000    0.009    0.000 tree.py:540(scan)
    11893    0.004    0.000    0.004    0.000 {built-in method isinstance}
      177    0.003    0.000    0.013    0.000 fake.py:47(search_scope)
        1    0.001    0.001    0.043    0.043 __init__.py:153(completions)
     4523    0.001    0.000    0.001    0.000 tree.py:42(is_node)
      170    0.001    0.000    0.002    0.000 copy.py:272(_reconstruct)
      185    0.001    0.000    0.014    0.000 classes.py:58(__init__)
   119/17    0.001    0.000    0.005    0.000 helpers.py:16(copy_node)
   267/29    0.001    0.000    0.016    0.001 cache.py:21(wrapper)
1496/1456    0.001    0.000    0.002    0.000 {built-in method getattr}
     4205    0.001    0.000    0.001    0.000 tree.py:265(__str__)
  721/716    0.001    0.000    0.020    0.000 cache.py:139(wrapper)
       18    0.001    0.000    0.001    0.000 {built-in method dir}
      722    0.001    0.000    0.001    0.000 classes.py:75(name)
  379/355    0.001    0.000    0.009    0.000 tree.py:112(get_parent_until)
      171    0.001    0.000    0.004    0.000 copy.py:67(copy)
  152/150    0.001    0.000    0.002    0.000 cache.py:154(wrapper)
      255    0.001    0.000    0.001    0.000 helpers.py:149(__init__)
      160    0.001    0.000    0.017    0.000 __init__.py:300(parent)
      170    0.001    0.000    0.001    0.000 {method '__reduce_ex__' of 'object' objects}
     4128    0.001    0.000    0.001    0.000 {method 'append' of 'list' objects}
      185    0.001    0.000    0.001    0.000 classes.py:378(_complete)
  543/511    0.000    0.000    0.017    0.000 _compatibility.py:124(<lambda>)
       35    0.000    0.000    0.000    0.000 {method 'search' of '_sre.SRE_Pattern' objects}
      160    0.000    0.000    0.016    0.000 __init__.py:456(_create_from_name)
     2100    0.000    0.000    0.000    0.000 tree.py:649(name)
      161    0.000    0.000    0.016    0.000 fake.py:100(get_faked)
      179    0.000    0.000    0.001    0.000 __init__.py:248(<lambda>)
      191    0.000    0.000    0.000    0.000 {built-in method hasattr}
      179    0.000    0.000    0.009    0.000 tree.py:526(subscopes)
      234    0.000    0.000    0.000    0.000 tree.py:226(__repr__)
      228    0.000    0.000    0.001    0.000 __init__.py:100(wrap)
      185    0.000    0.000    0.014    0.000 classes.py:368(__init__)
      611    0.000    0.000    0.000    0.000 _compatibility.py:120(__get__)
       17    0.000    0.000    0.000    0.000 inspect.py:484(cleandoc)
       94    0.000    0.000    0.001    0.000 tree.py:491(__repr__)
      960    0.000    0.000    0.000    0.000 tree.py:775(name)
      301    0.000    0.000    0.000    0.000 tree.py:151(__init__)
    34/17    0.000    0.000    0.017    0.001 __init__.py:285(execute)
      769    0.000    0.000    0.000    0.000 {built-in method setattr}
      143    0.000    0.000    0.001    0.000 __init__.py:259(__getitem__)
      161    0.000    0.000    0.014    0.000 fake.py:70(_faked)
       34    0.000    0.000    0.001    0.000 stdlib.py:27(execute)
      214    0.000    0.000    0.009    0.000 tree.py:538(_search_in_scope)
        1    0.000    0.000    0.002    0.002 __init__.py:267(values)
      160    0.000    0.000    0.001    0.000 __init__.py:285(__init__)
        1    0.000    0.000    0.001    0.001 {built-in method sorted}
       51    0.000    0.000    0.001    0.000 representation.py:261(get_instance_el)
      162    0.000    0.000    0.000    0.000 fake.py:122(is_class_instance)
      322    0.000    0.000    0.000    0.000 inspect.py:185(isgenerator)
      746    0.000    0.000    0.000    0.000 {method 'startswith' of 'str' objects}
       37    0.000    0.000    0.000    0.000 {method 'sub' of '_sre.SRE_Pattern' objects}
  162/128    0.000    0.000    0.001    0.000 common.py:49(wrapper)
    58/23    0.000    0.000    0.016    0.001 cache.py:56(__call__)
1052/1044    0.000    0.000    0.000    0.000 {built-in method len}
       17    0.000    0.000    0.001    0.000 _compatibility.py:163(literal_eval)
      185    0.000    0.000    0.000    0.000 classes.py:201(in_builtin_module)
     10/6    0.000    0.000    0.000    0.000 sre_parse.py:441(_parse)
       17    0.000    0.000    0.004    0.000 representation.py:595(get_return_types)
      138    0.000    0.000    0.000    0.000 __init__.py:44(__init__)
      549    0.000    0.000    0.000    0.000 {method 'lower' of 'str' objects}
      185    0.000    0.000    0.001    0.000 classes.py:396(complete)
       17    0.000    0.000    0.010    0.001 representation.py:555(py__call__)
       17    0.000    0.000    0.016    0.001 representation.py:78(__init__)
       17    0.000    0.000    0.005    0.000 representation.py:588(__init__)
      187    0.000    0.000    0.006    0.000 helpers.py:158(get_definition)
      170    0.000    0.000    0.000    0.000 copyreg.py:87(__newobj__)
       17    0.000    0.000    0.016    0.001 __init__.py:50(actual)
       17    0.000    0.000    0.004    0.000 __init__.py:164(get_subscope_by_name)
       17    0.000    0.000    0.000    0.000 recursion.py:132(push_execution)
      198    0.000    0.000    0.000    0.000 inspect.py:77(isclass)
    68/25    0.000    0.000    0.000    0.000 {method 'join' of 'str' objects}
       34    0.000    0.000    0.000    0.000 __init__.py:201(name)
      146    0.000    0.000    0.001    0.000 {built-in method repr}
      161    0.000    0.000    0.000    0.000 fake.py:18(_load_faked_module)
       24    0.000    0.000    0.001    0.000 re.py:277(_compile)
       17    0.000    0.000    0.003    0.000 docstrings.py:186(find_return_types)
       17    0.000    0.000    0.001    0.000 tree.py:63(raw_doc)
       17    0.000    0.000    0.005    0.000 helpers.py:7(deep_ast_copy)
       55    0.000    0.000    0.000    0.000 sre_parse.py:195(__next)
      170    0.000    0.000    0.000    0.000 {built-in method issubclass}
     16/5    0.000    0.000    0.000    0.000 sre_compile.py:70(_compile)
    68/17    0.000    0.000    0.010    0.001 _compatibility.py:122(<lambda>)
   206/28    0.000    0.000    0.000    0.000 tree.py:432(<genexpr>)
       17    0.000    0.000    0.000    0.000 representation.py:331(name)
      146    0.000    0.000    0.000    0.000 inspect.py:68(ismodule)
        4    0.000    0.000    0.000    0.000 {built-in method print}
    34/17    0.000    0.000    0.017    0.001 debug.py:48(wrapper)
     20/9    0.000    0.000    0.000    0.000 sre_parse.py:152(getwidth)
        1    0.000    0.000    0.044    0.044 {built-in method exec}
      612    0.000    0.000    0.000    0.000 {method 'items' of 'dict' objects}
        6    0.000    0.000    0.000    0.000 tokenize.py:157(generate_tokens)
      358    0.000    0.000    0.000    0.000 {method 'get' of 'dict' objects}
       17    0.000    0.000    0.000    0.000 textwrap.py:400(dedent)
       17    0.000    0.000    0.000    0.000 representation.py:293(parent)
        2    0.000    0.000    0.002    0.001 finder.py:51(filter_definition_names)
       17    0.000    0.000    0.000    0.000 ast.py:38(literal_eval)
       17    0.000    0.000    0.017    0.001 representation.py:744(parent_callback)
       17    0.000    0.000    0.000    0.000 {method 'findall' of '_sre.SRE_Pattern' objects}
       17    0.000    0.000    0.004    0.000 recursion.py:95(run)
       84    0.000    0.000    0.001    0.000 representation.py:562(__getattr__)
      172    0.000    0.000    0.000    0.000 {built-in method min}
        2    0.000    0.000    0.000    0.000 finder.py:63(<listcomp>)
       17    0.000    0.000    0.000    0.000 representation.py:306(get_parent_until)
        1    0.000    0.000    0.004    0.004 __init__.py:252(_simple_complete)
      131    0.000    0.000    0.000    0.000 tree.py:197(get_code)
       17    0.000    0.000    0.000    0.000 {method 'index' of 'list' objects}
       17    0.000    0.000    0.010    0.001 representation.py:358(py__call__)
     49/6    0.000    0.000    0.000    0.000 tree.py:431(get_code)
        4    0.000    0.000    0.000    0.000 copyreg.py:96(_slotnames)
       33    0.000    0.000    0.000    0.000 keywords.py:38(__init__)
      170    0.000    0.000    0.000    0.000 {built-in method __new__ of type object at 0x7f5843c220c0}
       17    0.000    0.000    0.000    0.000 representation.py:66(get_parent_until)
        4    0.000    0.000    0.000    0.000 debug.py:75(speed)
       50    0.000    0.000    0.000    0.000 tree.py:158(start_pos)
       17    0.000    0.000    0.000    0.000 __init__.py:495(wrapper)
      221    0.000    0.000    0.000    0.000 {built-in method id}
       43    0.000    0.000    0.000    0.000 representation.py:69(parent)
       17    0.000    0.000    0.001    0.000 representation.py:676(returns)
       34    0.000    0.000    0.000    0.000 representation.py:58(__init__)
       34    0.000    0.000    0.000    0.000 representation.py:287(__init__)
      152    0.000    0.000    0.000    0.000 {method 'setdefault' of 'dict' objects}
       20    0.000    0.000    0.000    0.000 {method 'split' of 'str' objects}
       17    0.000    0.000    0.000    0.000 representation.py:312(get_decorated_func)
      320    0.000    0.000    0.000    0.000 __init__.py:306(parent)
       46    0.000    0.000    0.000    0.000 sre_parse.py:142(__getitem__)
       11    0.000    0.000    0.000    0.000 {method 'match' of '_sre.SRE_Pattern' objects}
       17    0.000    0.000    0.004    0.000 representation.py:167(get_subscope_by_name)
       22    0.000    0.000    0.000    0.000 tree.py:479(__init__)
        6    0.000    0.000    0.000    0.000 tree.py:466(__repr__)
       17    0.000    0.000    0.000    0.000 recursion.py:128(pop_execution)
       68    0.000    0.000    0.001    0.000 representation.py:346(__getattr__)
      8/5    0.000    0.000    0.000    0.000 sre_parse.py:363(_parse_sub)
        6    0.000    0.000    0.000    0.000 representation.py:384(name)
        5    0.000    0.000    0.001    0.000 sre_compile.py:563(compile)
       17    0.000    0.000    0.001    0.000 tree.py:779(yields)
        5    0.000    0.000    0.000    0.000 sre_parse.py:750(parse)
        1    0.000    0.000    0.000    0.000 keywords.py:23(<listcomp>)
       18    0.000    0.000    0.000    0.000 __init__.py:123(type)
       17    0.000    0.000    0.001    0.000 docstrings.py:188(search_return_in_docstr)
        1    0.000    0.000    0.044    0.044 <string>:1(<module>)
       17    0.000    0.000    0.001    0.000 tree.py:784(is_generator)
        1    0.000    0.000    0.000    0.000 __init__.py:93(__init__)
        1    0.000    0.000    0.000    0.000 fast.py:358(_parse)
       69    0.000    0.000    0.000    0.000 debug.py:59(dbg)
        3    0.000    0.000    0.000    0.000 user_context.py:71(_calc_path_until_cursor)
        5    0.000    0.000    0.000    0.000 sre_compile.py:444(_compile_info)
       43    0.000    0.000    0.000    0.000 sre_parse.py:220(get)
        4    0.000    0.000    0.000    0.000 sre_compile.py:255(_optimize_charset)
      172    0.000    0.000    0.000    0.000 {method 'lstrip' of 'str' objects}
       24    0.000    0.000    0.000    0.000 tree.py:405(__init__)
       12    0.000    0.000    0.000    0.000 user_context.py:40(_backwards_line_generator)
       51    0.000    0.000    0.000    0.000 helpers.py:55(<listcomp>)
       20    0.000    0.000    0.000    0.000 re.py:172(sub)
       17    0.000    0.000    0.000    0.000 docstrings.py:116(_evaluate_for_statement_string)
       17    0.000    0.000    0.000    0.000 {method 'expandtabs' of 'str' objects}
       34    0.000    0.000    0.000    0.000 representation.py:349(isinstance)
       16    0.000    0.000    0.000    0.000 user_context.py:249(get_line)
        1    0.000    0.000    0.000    0.000 __init__.py:208(convert_leaf)
      148    0.000    0.000    0.000    0.000 __init__.py:297(is_definition)
        1    0.000    0.000    0.000    0.000 fast.py:419(_get_node)
        1    0.000    0.000    0.000    0.000 __init__.py:94(__init__)
        2    0.000    0.000    0.000    0.000 fast.py:114(__init__)
       17    0.000    0.000    0.000    0.000 __init__.py:48(py__call__)
       17    0.000    0.000    0.000    0.000 ast.py:49(_convert)
        1    0.000    0.000    0.004    0.004 __init__.py:161(get_completions)
       17    0.000    0.000    0.000    0.000 param.py:16(__init__)
       25    0.000    0.000    0.000    0.000 sre_parse.py:150(append)
       17    0.000    0.000    0.000    0.000 ast.py:30(parse)
        1    0.000    0.000    0.000    0.000 __init__.py:83(__init__)
       17    0.000    0.000    0.000    0.000 tree.py:380(__eq__)
        2    0.000    0.000    0.000    0.000 fast.py:263(_split_parts)
       31    0.000    0.000    0.000    0.000 sre_parse.py:214(match)
       17    0.000    0.000    0.000    0.000 <frozen importlib._bootstrap>:2264(_handle_fromlist)
       17    0.000    0.000    0.000    0.000 tree.py:358(__eq__)
        5    0.000    0.000    0.000    0.000 representation.py:715(names_dicts)
        4    0.000    0.000    0.000    0.000 user_context.py:53(_get_backwards_tokenizer)
     18/7    0.000    0.000    0.000    0.000 {built-in method next}
        3    0.000    0.000    0.000    0.000 tree.py:234(end_pos)
       17    0.000    0.000    0.017    0.001 helpers.py:173(parent)
        3    0.000    0.000    0.000    0.000 user_context.py:217(get_context)
        6    0.000    0.000    0.000    0.000 finder.py:443(global_names_dict_generator)
        5    0.000    0.000    0.000    0.000 sre_compile.py:548(_code)
        1    0.000    0.000    0.001    0.001 user_context.py:279(_parser)
       16    0.000    0.000    0.000    0.000 sre_parse.py:90(__init__)
       22    0.000    0.000    0.000    0.000 sre_parse.py:138(__len__)
        6    0.000    0.000    0.000    0.000 representation.py:487(__init__)
        1    0.000    0.000    0.000    0.000 representation.py:742(_module_attributes_dict)
        4    0.000    0.000    0.000    0.000 sre_compile.py:227(_compile_charset)
       12    0.000    0.000    0.000    0.000 tree.py:423(start_pos)
        2    0.000    0.000    0.000    0.000 finder.py:34(filter_after_position)
        5    0.000    0.000    0.000    0.000 sre_parse.py:190(__init__)
       17    0.000    0.000    0.000    0.000 tree.py:109(isinstance)
        1    0.000    0.000    0.000    0.000 parse.py:187(pop)
        1    0.000    0.000    0.000    0.000 fake.py:53(get_module)
        5    0.000    0.000    0.000    0.000 {built-in method compile}
        1    0.000    0.000    0.000    0.000 __init__.py:508(create)
        1    0.000    0.000    0.000    0.000 __init__.py:170(convert_node)
        4    0.000    0.000    0.000    0.000 sre_compile.py:418(_simple)
        3    0.000    0.000    0.000    0.000 user_context.py:161(get_line)
        1    0.000    0.000    0.000    0.000 parse.py:117(addtoken)
        9    0.000    0.000    0.000    0.000 _compatibility.py:142(u)
        1    0.000    0.000    0.000    0.000 __init__.py:136(_parsed_callback)
        5    0.000    0.000    0.000    0.000 representation.py:749(<genexpr>)
       12    0.000    0.000    0.000    0.000 user_context.py:56(<lambda>)
        4    0.000    0.000    0.000    0.000 tree.py:130(get_parent_scope)
        1    0.000    0.000    0.000    0.000 __init__.py:352(remove_last_newline)
        7    0.000    0.000    0.000    0.000 {method 'insert' of 'list' objects}
        6    0.000    0.000    0.000    0.000 {method 'replace' of 'str' objects}
        1    0.000    0.000    0.000    0.000 __init__.py:38(load_grammar)
        1    0.000    0.000    0.000    0.000 helpers.py:10(completion_parts)
      3/2    0.000    0.000    0.000    0.000 fast.py:184(parent_until_indent)
        1    0.000    0.000    0.000    0.000 posixpath.py:145(dirname)
       10    0.000    0.000    0.000    0.000 sre_compile.py:545(isstring)
        2    0.000    0.000    0.000    0.000 __init__.py:324(_tokenize)
        1    0.000    0.000    0.000    0.000 fast.py:451(__init__)
        4    0.000    0.000    0.000    0.000 helpers.py:169(__init__)
        2    0.000    0.000    0.000    0.000 fast.py:172(close)
        1    0.000    0.000    0.000    0.000 keywords.py:29(keyword_names)
        1    0.000    0.000    0.000    0.000 fast.py:24(__init__)
        1    0.000    0.000    0.000    0.000 representation.py:752(name)
        1    0.000    0.000    0.001    0.001 user_context.py:291(user_stmt)
        1    0.000    0.000    0.000    0.000 fast.py:245(_reset_caches)
        1    0.000    0.000    0.000    0.000 fast.py:100(__call__)
        4    0.000    0.000    0.000    0.000 sre_parse.py:307(_escape)
        3    0.000    0.000    0.000    0.000 common.py:60(__init__)
        1    0.000    0.000    0.000    0.000 cache.py:48(clear_time_caches)
        5    0.000    0.000    0.000    0.000 sre_parse.py:67(__init__)
        1    0.000    0.000    0.000    0.000 tree.py:446(get_statement_for_position)
        1    0.000    0.000    0.000    0.000 keywords.py:21(keywords)
        1    0.000    0.000    0.000    0.000 __init__.py:445(get_by_name)
        2    0.000    0.000    0.000    0.000 fast.py:31(reset_caches)
        2    0.000    0.000    0.000    0.000 re.py:192(split)
        1    0.000    0.000    0.000    0.000 {built-in method __import__}
        1    0.000    0.000    0.000    0.000 tokenize.py:150(source_tokens)
       17    0.000    0.000    0.000    0.000 {method 'endswith' of 'str' objects}
        2    0.000    0.000    0.000    0.000 tree.py:427(end_pos)
        1    0.000    0.000    0.000    0.000 fast.py:157(reset_node)
        1    0.000    0.000    0.000    0.000 fast.py:226(remove_last_newline)
        1    0.000    0.000    0.000    0.000 parse.py:63(__init__)
        2    0.000    0.000    0.000    0.000 {method 'split' of '_sre.SRE_Pattern' objects}
        1    0.000    0.000    0.000    0.000 fast.py:43(global_names)
        1    0.000    0.000    0.000    0.000 keywords.py:30(<listcomp>)
        1    0.000    0.000    0.000    0.000 helpers.py:37(check_error_statements)
        1    0.000    0.000    0.000    0.000 user_context.py:317(user_scope)
        1    0.000    0.000    0.000    0.000 fast.py:249(update)
        1    0.000    0.000    0.000    0.000 user_context.py:157(call_signature)
        6    0.000    0.000    0.000    0.000 {built-in method time}
        3    0.000    0.000    0.000    0.000 common.py:75(__next__)
…