@davidhalter 2015-11-19
Very very nice! Thank you for finding the issue!

Although I'm not sure about the fix. I think a "get_faked" call should be called only once. Couldn't we just add a `jedi.cache.memoize_method` in front of it? (Maybe we have to add a `cache.memoize_function`)

Maybe a regression test would be nice as well. If it's not easily possible with the `faked` API, just use the Jedi API, run it twice and inspect the `Definition._definition` object.