@ColinDuquesnoy 2015-11-19
> Couldn't we just add a jedi.cache.memoize_method in front of it? (Maybe we have to add a cache.memoize_function)

Sounds better indeed. I'll give it a try during the weekend.