@ColinDuquesnoy 2015-12-13
Hi @davidhalter 

I've added a regression test and a ``memoize_function`` to ``jedi.cache``. ``jedi.evaluate.compiled.fake.get_faked`` now uses that function. (the implementation comes straight from the [PythonDecoratorLibrary](https://wiki.python.org/moin/PythonDecoratorLibrary#Memoize), I've just removed the call to ``functools.wrap``).