@davidhalter 2016-01-05
@ColinDuquesnoy I'm not sure how to think about the PR right now, since the linter branch takes now a bit of a different approach in how it loads compiled objects. However this PR is still badly needed and I intend to merge, it probably just takes some time until things have sorted out.