@ColinDuquesnoy 2016-07-21
@davidhalter I've just checked the latest dev branch and issue #591 still happens. I've synced my fork and it looks like the fix still applies too. I remember I checked the linter branch a few months ago and the issue was gone... What's happening?