@davidhalter 2016-07-21
The issue was never really gone. It's just that the impact might have been smaller. I still have to find a way to get it in before 0.10.0.