@ColinDuquesnoy 2016-07-21
Thank you for the clarification. What about this PR? Do you want to merge it or do you think there is a better fix?