@davidhalter 2016-07-22
Well the fix __was__ good! But some of the the other logic has changed quite a bit and now I need to refactor your fix just a little bit.