@davidhalter 2016-07-31
All right! I have finally merged it. I did this with the intention of providing a fixup. You can read about the actual issue in the commit note. Thanks anyway! Especially tracking it down and providing a test helped a lot!