@davidhalter 2016-07-31
Oh wow, this one is actually huge. I think it brought down testing time by about 10s. Which means about 25% speed improvement, thanks again :)