@ColinDuquesnoy 2016-07-31
Wow. Glad I could help ;-)